adding the code from asheroto to install winget (#684)
* adding the code from asheroto to get winget

with Server versions of Windows, that come without winget , the script seems unable to install it and then refuses to install any programs. so by adding two lines of code (hopefully) you can include the installation of winget and then fulfill the requirement and let the program run and do all the cool things

* instaling winget from asheroto

so the script will check if winget is installed on server and LTSC versions of Windows.  and tries to install it, when for some reason it can't, then a condition on the code gets fulfilled and the function breaks, this code from the asheroto repo should, (hopefully) make it so winget can get installed and the condition is never met, therefore allowing it the script to continue