Improve handling of Adobe's CC Cleaner tool
Implement a try-catch-finally struct to do a better handling of Adobe's Creative Cloud Cleaner tool downloading and running of such program.

This will allow to see any errors that occurs if the URL changes or the tool fails to run.

I also have moved the cleaning up process to the finally part of the structure, as it makes the most sense.

This tries to fix the issue #1563.