Update inputXML.xaml (#2558)
Change to line 654 from {SearchBarTextBoxFontSize} to 12. This now now allows the text to display correctly in the text box.
Change to line 666 from 16 to 180 for the margin. This now allows text to be visible when searching, also positions the magnify glass to the right side of the text box.
Change to line 678 from 193 to 205, which now places the clear button (X) on the outside of the text box for clearing the text in the textbox.