[MicroWin] June-August 2025 Update (#3417)
* Disable "Microsoft account" notification source

* [WinPE] Remove driver addition block for WinPE

WE DO NOT TOUCH THE DRIVERS IN WinPE. There are reports of people getting "Install driver to show hardware" screens all over this repository, and on Discord; and the less drivers we touch in WinPE, the better.

Drivers can still be added to Preinstallation Environments in the following ways:

- Using the driver installation screens
- Firing up "drvload.exe <driver>" in cmd

* [Fix] Added fallback for DISM export command

This is a port of the fix in #3305

* [Cleanup] Removed some comments that no longer make sense

* [Fix] Same DISM export image fallback fix

* Merge branch 'main' into microwin-202506

* [Fix] Improve UI consistency for instructions

Fixes #3394

* Merge branch 'main' into microwin-202506

* [Unattended answer file] Remove it from drive root

The answer file, on the drive root, is not necessary for us to apply it. In fact, it's not even used there

* Merge branch 'main' into microwin-202506

* [MicroWin] June-August 2025 Update (#3) -- Contributions from Callum

* Allow people without compatible hardware or a USB to use MicroWin.

* Update functions/microwin/Microwin-NewUnattend.ps1

Co-authored-by: CodingWonders <101426328+CodingWonders@users.noreply.github.com>

* Update Invoke-Microwin.ps1

* Update Microwin-NewUnattend.ps1

* Update Microwin-NewUnattend.ps1

* Add error pop up if ISO Creation fails.

Issue 2653

* Add Disable WPBT Execution to MicroWin.

* Update functions/microwin/Invoke-Microwin.ps1

Co-authored-by: CodingWonders <101426328+CodingWonders@users.noreply.github.com>

* modified:   functions/microwin/Invoke-Microwin.ps1
	modified:   xaml/inputXML.xaml

---------

Co-authored-by: CodingWonders <101426328+CodingWonders@users.noreply.github.com>

* Add conversion to ESD (#4)

* Add conversion to ESD

Issue - #3450

* Update Invoke-Microwin.ps1

Added quotes to the file paths. Put all the arguments in 1 string (as that also works fine)

---------

Co-authored-by: CodingWonders <101426328+CodingWonders@users.noreply.github.com>

* Update MicroWin contributor list

* Merge branch 'main' into microwin-202506

* Merge branch 'main' into microwin-202506

* [MicroWin] Add automatic configuration settings

Originally implemented in #2618. Adapted to follow the new file structure. And it works.

Though there are issues that will be detailed very soon

* [Fix] Fixed typos, updated descriptions

* Re-add WinPE driver addition

We're not yet sure if that is the actual problem of missing storage controllers. Logs can tell us more about this. Maybe for a future PR?

* [Fix/WPBT] Add spaces to reg key path

Avoid REG failure

* [Fix/XAML] Fix word wrapping issue for checkboxes